Left-aligning the main elements
It's now a border that contains an image, instead of an ellipse. This lets them move together. The idea is that a long-press or a right-click will pop up the menu to the right of the button, so the window needs to expand horizontally. The items shouldn't center-align or shift as the window expands.